server/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sqlalchemy-utils==0.41.1
asyncio==3.4.3
aiohttp==3.9.1